requirements.txt
blinker==1.8.2
click==8.1.7
dotenv-python==0.0.1
Flask==3.0.3
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
packaging==24.1
peewee==3.17.7
python-dotenv==1.0.1
Werkzeug==3.0.4
